// src/app/components/interfaces/orders.interfaces.ts
export interface users {
  name_complet: string;
}
export interface Orders {
  nombre_cliente    : string;
  sucursal          : string;
  sem_entrega       : string;
  fecha_entrega     :Date;
  folio             :string;
  num_orden_com     :string;
  ordenpro          :string;
  avance            :string;
  stock             :number;
}

export interface infoClient{
  city: String;
  plant: String;
}
export interface infoPlant{
  name_client:string;
  city: String;
  plant: String[];
  plant2:String;
}
export interface infoPart {
  plant:        string;
  no_parte:     string;
  descrip:      string;
  no_orden:     string;
  clave:        String;
  medida:       string;
  pieza_medida: number;
  articulo:     number;
  pieza_caja:   number;
}

export interface infoInputs {
  plant:        string;
  no_parte:     string;
  descrip:      string;
  no_orden:     string;
  clave:        String;
  medida:       string;
  pieza_medida: number;
  articulo:     number;
  pieza_caja:   number;
  num_visual:   string;
  autorizo:     string;
  stock:number;

}

// Generated by https://quicktype.io

export interface OrdersProductions {
  id_order:        string;
  folio?:          string;
  nombre_cliente?: string;
  ciudad?:         string;
  sucursal?:       string;
  ordenpro?:       string;
  fecha_entrega?:  string;
  fecha_creacion?: string;
  sem_entrega?:    number;
  num_articulo?:   number;
  num_parte?:      string;
  descripcion?:    string;
  num_orden_com?:  string;
  clave?:          string;
  medida?:         string;
  pieza_medida?:   number;
  hojas_cortar?:   number;
  num_paquetes?:   string;
  cantidad?:       number;
  piezas_caja?:    number;
  cajas_utilizar?: number;
  elaboro?:        string;
  autorizo?:       string;
  stock?:number;
}
export interface OrdersProductions2 {
  cantidad:       number;
}

export interface infoOrdersProductions {
  id_order:        number;
  num_articulo?:   number;
  num_parte?:      string;
  descripcion?:    string;
  num_orden_com?:  string;
  clave?:          string;
  medida?:         string;
  pieza_medida?:   number;
  hojas_cortar?:   number;
  num_paquetes?:   string;
  cantidad?:       number;
  piezas_caja?:    number;
  cajas_utilizar?: number;
  stock?:number;
}
export interface infosum {
  "SUM(num_paquetes)": number;
  "SUM(cajas_utilizar)":number;
}

export interface date{
  no_orden?:    string;
  total?:       number;
  peso_bruto?:  number;
  peso_neto?:   number;
  total_cajas?: number;
}



//bitacoras

export interface bitcorte{
  id_corte      :number;
  ordenpro      :string;
  operacion    ?:string;
  articulo     ?:number;
  velocidad    ?:number;
  observacion  ?:string;
  fecha        ?:string;
  hora         ?:string;
  responsable   :string;
  campo1       ?:string;
  campo2       ?:string;
  campo3       ?:string;
  campo4       ?:string;
  campo5       ?:string;
  campo6       ?:string;
  campo7       ?:string;
  campo8       ?:string;
  campo9       ?:string;
  campo10       ?:string;
  campo11       ?:string;
  campo12       ?:string;
  campo13       ?:string;
  campo14       ?:string;
  campo15       ?:string;
}

export interface bitsuaje{
  id_suaje      :number;
  ordenpro      :string;
  operacion    ?:string;
  posicion     ?:number;
  observacion  ?:string;
  fecha        ?:string;
  hora         ?:string;
  responsable   :string;
  campo1       ?:string;
  campo2       ?:string;
  campo3       ?:string;
  campo4       ?:string;
  campo5       ?:string;
  campo6       ?:string;
  campo7       ?:string;
  campo8       ?:string;
}

export interface bitlami{
  id_lami       :number;
  ordenpro      :string;
  operacion    ?:string;
  velocidadON  ?:number;
  velocidadOFF ?:number;
  fecha        ?:string;
  hora         ?:string;
  responsable   :string;
  campo1       ?:string;
  campo2       ?:string;
  campo3       ?:string;
  campo4       ?:string;
  campo5       ?:string;
  campo6       ?:string;
  campo7       ?:string;
  campo8       ?:string;
  observaciones       ?:string;

}


/* export interface selectorden{
ordenpro: String;
} */


// Generated by https://quicktype.io

export interface selectorden {
  id_order:       number;
  folio:          string;
  nombre_cliente: string;
  ciudad:         string;
  sucursal:       string;
  ordenpro:       string;
  fecha_entrega:  string;
  fecha_creacion: Date;
  sem_entrega:    number;
  num_articulo:   number;
  num_parte:      string;
  descripcion:    string;
  num_orden_com:  string;
  clave:          string;
  medida:         string;
  pieza_medida:   number;
  hojas_cortar:   number;
  num_paquetes:   string;
  cantidad:       number;
  piezas_caja:    number;
  cajas_utilizar: number;
  elaboro:        string;
  autorizo:       string;
}

export interface selectordenCal {
  id: number;
  no_orden      :string;
  total         :number;
  peso_bruto    :number;
  peso_neto     :number;
  total_cajas   :number;
}

export interface inspection{
  id:           number;
  ordenpro2     :string;
  num_parte     :string;
  descripcion   :string;
  medida        :string;
  modelo        :string;
  num_visual    :string;
  inicial       :number;
  dato1         :number;
  dato2         :number;
  dato3         :number;
  dato4         :number;
  dato5         :number;
  dato6         :number;
  dato7         :number;
  dato8         :number;
  dato9         :number;
  dato10        :number;
  observaciones :string;
  fecha_creacion:string;
  elaboro:string;
}

export interface papel{
id_papel:number;
ordenpro :string;
tipo     :string;
lote     :string;
elaboro  :string;
modelo   :string;
cantidad :number;
fecha_creacion :string;
}
export interface folio_inspect{
  id_folio:number;
  ordenpro :string;
  tipo     :string;
  lote     :string;
  elaboro  :string;
  modelo   :string;
  cantidad :number;
  fecha_creacion :string;
  }
export interface vigilar{
id_vigilar:number;
ordenpro       :string;
op1            :string;
op2            :string;
op3            :string;
op4            :string;
op5            :string;
cajas_apro     :number;
cajas_def      :number;
total          :number;
observaciones  :string;
elaboro        :string;
fecha_creacion :string;
}

export interface control{
  id_liberacion:number;
  ordenpro2    :string;
  cliente      :string;
  sucursal     :string;
  num_parte    :string;
  cantidad     :number;
  codigo       :string;
  muestra      :number
  piezasA      :number;
  piezasR      :number;
  total        :number;
  observaciones:string;
  elaboro      :string;
  nombre_cliente:string;
  AC           :string;
  op1          :string;
  op2          :string;
  fecha_ini    :string;
}

export interface salida{
  id_salida: number;
  ordenpro     :string;
  cliente      :string;
  sucursal     :string;
  num_parte    :string;
  cantidad     :number;
  codigo       :string;
  muestra      :number
  piezasA      :number;
  piezasR      :number;
  total        :number;
  observaciones:string;
  elaboro      :string;
  nombre_cliente:string;
  num_ac           :string;
  op1          :string;
  op2          :string;
  fecha_ini:string;
  inspection:string;
  fecha:string;
}

export interface avance {
  "COUNT(num_parte)": number;
  avance:number;
  prueba:number;
}

export interface avance2 {
  avance:number;
}


export interface stock {
  stock:number;

}

export interface observaciones {

  observaciones: string;
}

export interface AdminStock{
  plant       :string;
  no_parte    :string;
  descrip     :string;
  stock       :number;
  observaciones :string;

}

export interface list_inspectio{
  id_list_inspection       :number;
  ordenpro :string;
  motivo   :string;
  elabora  :string;
  cliente  :string;
  fecha    :string;
  identificacion :string;
  lote     :string;
  observaciones  :string;
  numero   :string;
  inspection :string;
  cumple :string;
  comentarios :string;
}